IOS-7443 CR fixes
