+++ Tangem/Modules/Send/Summary/Interactor/CommonSendTransactionSummaryDescriptionBuilderProtocol.swift
@@ -1,0 +1,42 @@
+//
+//  SendTransactionSummaryDescriptionBuilderProtocol.swift
+//  Tangem
+//
+//  Created by Aleksei Muraveinik on 25.07.24.
+//  Copyright © 2024 Tangem AG. All rights reserved.
+//
+
+import Foundation
+
+struct CommonSendTransactionSummaryDescriptionBuilder {
+    private let tokenItem: TokenItem
+    private let feeTokenItem: TokenItem
+
+    init(tokenItem: TokenItem, feeTokenItem: TokenItem) {
+        self.tokenItem = tokenItem
+        self.feeTokenItem = feeTokenItem
+    }
+}
+
+// MARK: - SendTransactionSummaryDescriptionBuilder
+
+extension CommonSendTransactionSummaryDescriptionBuilder: SendTransactionSummaryDescriptionBuilder {
+    func makeDescription(amount: Decimal, fee: Decimal) -> String? {
+        let amountInFiat = tokenItem.id.flatMap { BalanceConverter().convertToFiat(amount, currencyId: $0) }
+        let feeInFiat = feeTokenItem.id.flatMap { BalanceConverter().convertToFiat(fee, currencyId: $0) }
+        let totalInFiat = [amountInFiat, feeInFiat].compactMap { $0 }.reduce(0, +)
+
+        let formattingOptions = BalanceFormattingOptions(
+            minFractionDigits: BalanceFormattingOptions.defaultFiatFormattingOptions.minFractionDigits,
+            maxFractionDigits: BalanceFormattingOptions.defaultFiatFormattingOptions.maxFractionDigits,
+            formatEpsilonAsLowestRepresentableValue: true,
+            roundingType: BalanceFormattingOptions.defaultFiatFormattingOptions.roundingType
+        )
+
+        let formatter = BalanceFormatter()
+        let totalInFiatFormatted = formatter.formatFiatBalance(totalInFiat, formattingOptions: formattingOptions)
+        let feeInFiatFormatted = formatter.formatFiatBalance(feeInFiat, formattingOptions: formattingOptions)
+
+        return Localization.sendSummaryTransactionDescription(totalInFiatFormatted, feeInFiatFormatted)
+    }
+}
