+++ Tangem/Modules/Send/Summary/Interactor/SendSummaryInteractor.swift
@@ -20,36 +20,25 @@ class CommonSendSummaryInteractor {
 
     private let sendTransactionDispatcher: SendTransactionDispatcher
     private let descriptionBuilder: SendTransactionSummaryDescriptionBuilder
-    private let blockchain: Blockchain
 
     init(
         input: SendSummaryInput,
         output: SendSummaryOutput,
         sendTransactionDispatcher: SendTransactionDispatcher,
-        descriptionBuilder: SendTransactionSummaryDescriptionBuilder,
-        blockchain: Blockchain
+        descriptionBuilder: SendTransactionSummaryDescriptionBuilder
     ) {
         self.input = input
         self.output = output
         self.sendTransactionDispatcher = sendTransactionDispatcher
         self.descriptionBuilder = descriptionBuilder
-        self.blockchain = blockchain
     }
 
     private func mapToDescription(transaction: SendTransactionType) -> String? {
         switch transaction {
         case .transfer(let bsdkTransaction):
-            let isNoFiatFee = switch blockchain.feePaidCurrency {
-            case .feeResource:
-                true
-            default:
-                false
-            }
-
             return descriptionBuilder.makeDescription(
                 amount: bsdkTransaction.amount.value,
-                fee: bsdkTransaction.fee.amount.value,
-                isNoFiatFee: isNoFiatFee
+                fee: bsdkTransaction.fee.amount.value
             )
         case .staking(let stakingTransaction):
             return nil // Waiting texts
